requirements.txt
Flask==3.0.3
Flask-SocketIO==5.3.6
eventlet==0.36.1
gunicorn==23.0.0
python-dotenv==1.0.1
